Rename the module to `chsesame2` from `lock`
So far, we have used the abstract name, `lock`.
However, as there are plans to support devices other than **lock**s,
we should move to more declarative module names.
